started redesign into hexagonal pattern
The separation between Discord specific logic and domain logic caused
problems with the previous design attempt. Trying to put all Discord
specific logic in adapters. These adapters will call ports which are
Discord agnostic. The ports will then in turn call domain logic.

Example flow of a command interaction coming into application:

    [ Discord API ]
        |
        |
    Interaction
        |
        V
    [ Discord Adapter ]
        |
        |
         (calls)
                |
                |
            Domain representation of parameters
                |
                V
            [ Domain Logic Port ]
                |
                |
                 (calls)
                        |
                        |
                    Domain representation of parameters
                        |
                        V
                [ Domain Logic ]
                        |
                        |
                (returns)
                |
                |
            Agnostic representation of result
                |
                |
        (returns)
        |
        |
    Agnostic representation of result
        |
        V
    [ Discord Adapter ]
        |
        |
    Discord Interaction API Call
        |
        V
    [ Discord API ]

    ( This structure will probably change as I implement )

This will allow for each area of the architecture to be more clear,
and less cluttered by irrelevant details, or result in changing
business logic having a large blast radius.

Currently I have started to move the logic from src/bot/index.ts into
src/bot/commands/adapters/. In adapters/ the adapter.ts is the general
Discord adapter. The commands.ts file holds an adapter specifically
for Discord commands. This adapter will be called by the general
Discord adapter (as we need to receive Discord command interaction
events we need to register an interaction event handler. So it makes
more sense to share the general interaction handler with the general
Discord adapter. Then have a different adapter for the different types
of Discord interactions. Currently on the command sub-adapter is
implemented but the new Discord buttons and forms will probably lead
to other adapters).

The src/bot/commands directory should probably be moved out of the
bot/ directory. This bot/ directory should be deleted. A new
orchestration entry point will need to be written which instantiates
all the adapters.

A way of fitting src/bot/emojis (and assets/custom-emojis) into this design pattern must be
considered. This may relate to the later "[ Discord Adapter ]" that is
shown in the diagram above. This adapter will be responsible for
taking the domain representation of a result and transforming it into
a Discord API request to respond to a Discord interaction.

In [1] a driving side and driven side is discussed. The driving side
is Discord interaction events being received, and the driven side is
responses to those interaction events. It is possible that ports may
need to be written between the "[ Domain Logic ]" and the latter "[
Discord Adapter ]". However for this to occur the domain logic may
have to know to call this new port. And I'm not sure how to make
domain logic call ports, which then call adapters given this
model. Perhaps each "Domain representation of parameters" should
include an ID which can be associated back to the interaction so it
can be responded to?

[1] https://medium.com/ssense-tech/hexagonal-architecture-there-are-always-two-sides-to-every-story-bc0780ed7d9c