scripts: dev.sh: simplified shell action parsing, can elide CMD
Now the shell argument action parsing is way simpler and robust. It
uses regular expressions to grab parts of the action spec, instead of
a bespoke hand parser using awk and cut. As an improvement while
re-implementing the shell action, the CMD portion can now be elided,
as long as an equals sign is still there.

Example:
- "shell=" will execute "bash" in the "bot" container
- "shell=postgres=" will execute "bash" in the "postgres" container