refactored so config was an interface
The Config data structure and its sub-structures were classes
initially. This was slightly undesirable from a typing standpoint as
the actual fields of the Config structure were the only thing that
mattered.

Additionally by having a class be the main data structure for Config
it tied Config up to always load from the environment. Making testing
potentially troublesome.

Config was made into a class so the envVar decorator could be used on
the fields. But now separate implementation classes (EnvConfig,
EnvDiscordConfig, EnvPostgresConfig) have those decorators.