reworked config pattern to have working types
The previous Config typing system w the @envVar() annotation was not
possible. Due to the fact that Typescript conditional types cannot be
used with type guards to narrow a generic type parameter and return
the correct type.

Ex - This won't work:

```ts
type ABCType = "return-a-string" | "return-a-number";

function foo<T extends ABCType>(value: T): T
extends "return-a-string" ? string : T extends "return-a-number" ?
number : never {
  if (isReturnAString(value)) {
    return "hey I know I should return a string"; // <-- ERROR
  } else if (isReturnANumber(value)) {
    return 314; // <-- ERROR
  }
}

function isReturnAString(val: ABCType): val is "return-a-string" {
  return val === "return-a-string";
}

function isReturnANumber(val: ABCType): val is "return-a-number" {
  return val === "return-a-number";
}
```

The problem is that Typescript only allows one to do type logic (ie.,
conditional types) at compile time. So if a conditional type cannot be
resolved at compile time then it is not possible.

This is why the above `foo()` function does not work. The generic type
parameter is not known to be type `"return-a-string"` or type
`"return-a-number"` until runtime. Where type guards
`isReturnAString()` and `isReturnANumber()` are used.

The Config `@envVar()` annotation relied heavily on this type of
conditional type at runtime logic to know what type the environment
variable should be serialized into.

The new solution leaves this to the caller when assigning a field. So
instead of one function which returns any of the possible environment
variable types there are instead 3 functions which return the possible
types. Which one to call is up to the implementation which is using
the `@envVar()` annotation value.